+++ CHANGES.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+5.0.13 (unreleased)
+-------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
+
+
 5.0.12 (2016-09-09)
 -------------------